Refator logic in timer func
